+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Commands for mirroring elements.
+    - `Mirror Elements Across X Axis`
+    - `Mirror Elements Across Z Axis`
+    - `Mirror Elements Across Y Axis`
+
+### Changed
+
+- Reimplement paste logic to avoid overriding built-in paste action.
+
 ## [0.0.7] - 2025.03.06
 
 ### Added